Fix → changed parseFloat in ParseInt into Age prompt

+++ js/app.js
@@ -14,7 +14,7 @@ console.log("JS Trial")
 
 
 const distance = parseFloat(prompt("Inserisci i km che vuoi percorrere (solo il numero)")); //Number
-const age = parseFloat(prompt("Inserisci la tua età (solo il numero)")); //Number
+const age = parseInt(prompt("Inserisci la tua età (solo il numero)")); //Number
 
 // console.log("Distanza Percorsa " + distance)
 // console.log("Età Passeggero " + age)
